+++ src/api/recipes/imports/scrapes.ts
@@ -13,7 +13,7 @@ import {
 	transformScrapeToRecipe,
 	validateAsRecipe,
 } from '@/steps';
-import type { RecipeScrapeDBRead } from '@/types';
+import type { RecipeIngredientWriteInput, RecipeInstructionWriteInput, RecipeScrapeDBRead } from '@/types';
 
 export default {
 	post: [requireAuthentication, requirePermissions('recipes:scrape'), _postHandler] as const,
@@ -53,7 +53,7 @@ export async function _postHandler({ request, ctx }: RequestInfo<DefaultAppConte
 				const savedSection = savedSections[index];
 				return {
 					sectionId: savedSection.id,
-					instructions: section.instructions,
+					instructions: section.instructions as RecipeInstructionWriteInput[],
 				};
 			}),
 		);
@@ -65,7 +65,7 @@ export async function _postHandler({ request, ctx }: RequestInfo<DefaultAppConte
 				const savedSection = savedSections[index];
 				return {
 					sectionId: savedSection.id,
-					ingredients: section.ingredients,
+					ingredients: section.ingredients as RecipeIngredientWriteInput[],
 				};
 			}),
 		);
@@ -73,11 +73,29 @@ export async function _postHandler({ request, ctx }: RequestInfo<DefaultAppConte
 		await updateRecipeScrapeStatus(recipeScrape.id, 'INGREDIENTS_SAVED', null, userId, ctx.logger);
 
 		recipeScrape = await updateRecipeScrapeStatus(recipeScrape.id, 'COMPLETED', null, userId, ctx.logger);
-		await createRecipeScrapeAttempt(recipeScrape.id, 'api', null, 'COMPLETED', null, null, userId, ctx.logger);
+		await createRecipeScrapeAttempt(
+			recipeScrape.id,
+			'api',
+			null,
+			'COMPLETED',
+			null,
+			import.meta.env.VITE_APP_VERSION,
+			userId,
+			ctx.logger,
+		);
 	} catch (err) {
 		if (recipeScrape) {
 			await updateRecipeScrapeStatus(recipeScrape.id, 'FAILED', (err as Error).message, userId, ctx.logger);
-			await createRecipeScrapeAttempt(recipeScrape.id, 'api', null, 'FAILED', (err as Error).message, null, userId, ctx.logger);
+			await createRecipeScrapeAttempt(
+				recipeScrape.id,
+				'api',
+				null,
+				'FAILED',
+				(err as Error).message,
+				import.meta.env.VITE_APP_VERSION,
+				userId,
+				ctx.logger,
+			);
 		}
 		return apiErrorResponse(err, 'Error processing recipe scrape');
 	}