+++ src/pages/recipes/routes.ts
@@ -1,11 +1,12 @@
 import { route } from 'rwsdk/router';
 import Pages__recipes__edit from './edit';
 import Pages__recipes__listing from './listing';
+import Pages__recipes__print from './print';
 import Pages__recipes__upload from './upload';
 import Pages__recipes__uploadView from './upload-view';
 import Pages__recipes__view from './view';
 
-export default [
+const appRoutes = [
 	route('/', Pages__recipes__listing),
 	route('/uploads/new', Pages__recipes__upload),
 	route('/new', Pages__recipes__edit),
@@ -13,3 +14,10 @@ export default [
 	route('/:recipeId/edit', Pages__recipes__edit),
 	route('/uploads/:recipeUploadId', Pages__recipes__uploadView),
 ];
+
+const noJsRoutes = [route('/:recipeId/print', Pages__recipes__print)];
+
+export default {
+	app: appRoutes,
+	noJS: noJsRoutes,
+};