+++ src/pages/recipes/upload.tsx
@@ -1,11 +1,11 @@
 import type { RequestInfo } from 'rwsdk/worker';
 import RecipeUpload from '@/forms/recipe-upload';
-import StandardLayout from '@/layouts/standard';
+import AppLayout from '@/layouts/app';
 
 export default async function Pages__recipes__upload({ ctx }: RequestInfo): Promise<React.JSX.Element> {
 	return (
-		<StandardLayout currentBasePage="recipes" pageTitle="Recipes" ctx={ctx}>
+		<AppLayout currentBasePage="recipes" pageTitle="Recipes" ctx={ctx}>
 			<RecipeUpload />
-		</StandardLayout>
+		</AppLayout>
 	);
 }